api/requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python 
# azure-monitor-opentelemetry 

azure-functions
pandas
numpy
scikit-learn
lightgbm
joblib
pymongo
requests
python-dotenv